backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
redis==5.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pydantic[email]==2.11.1
python-dotenv==1.0.1
# ── AI Pipeline dependencies ──────────────────────────────────────────────────
groq==0.22.0
pandas==2.2.3
requests==2.32.3
